fix uid sid

+++ src/kvraft/common.go
@@ -16,8 +16,8 @@ type PutAppendArgs struct {
 	// You'll have to add definitions here.
 	// Field names must start with capital letters,
 	// otherwise RPC will break.
-	SID   int64
-	UID	  int
+	SID   int	// request id, increment from 0. 
+	UID	  int64 // client id, a random int64.
 }
 
 type PutAppendReply struct {
@@ -27,8 +27,8 @@ type PutAppendReply struct {
 type GetArgs struct {
 	Key string
 	// You'll have to add definitions here.
-	SID int64
-	UID int
+	SID int
+	UID int64
 }
 
 type GetReply struct {
